Debug log hypersphere TF

+++ geomstats/backend/numpy.py
@@ -4,7 +4,7 @@ import numpy as np
 
 int32 = np.int32
 int8 = np.int8
-double = np.float32
+float32 = np.float32
 float64 = np.float64
